tile masks and fix some bugs of walking

+++ include/Integration.hpp
@@ -21,6 +21,7 @@
 
 #define FONTPATH  		ASSETS "ref/font/Cubic_11_1.010_R.ttf"
 
+#include <unordered_set>
 #include "pch.hpp"
 #include "Util/Logger.hpp"
 #include "Util/GameObject.hpp"
@@ -30,8 +31,8 @@ namespace std {
     template <>
     struct hash<glm::ivec2> {
         std::size_t operator()(const glm::ivec2& v) const noexcept {
-            std::size_t h1 = std::hash<float>{}(v.x);
-            std::size_t h2 = std::hash<float>{}(v.y);
+            std::size_t h1 = std::hash<int>{}(v.x);
+            std::size_t h2 = std::hash<int>{}(v.y);
             return h1 ^ (h2 << 1);
         }
     };
